refactor: 쿠폰 다운로드 CouponService와 Controller 클래스로 이동
