+++ src/main/kotlin/com/mojh/cms/coupon/controller/CouponController.kt
@@ -2,6 +2,7 @@ package com.mojh.cms.coupon.controller
 
 import com.mojh.cms.common.ApiResponse
 import com.mojh.cms.coupon.dto.CreateCouponRequest
+import com.mojh.cms.coupon.dto.MemberCouponResponse
 import com.mojh.cms.coupon.service.CouponService
 import com.mojh.cms.member.entity.Member
 import com.mojh.cms.security.LoginMember
@@ -25,10 +26,16 @@ class CouponController(
 
     @PostMapping("/{couponId}/enable")
     @Secured("ROLE_SELLER")
-    fun enableCoupon(@PathVariable couponId: Long,
-                     @LoginMember seller: Member): ApiResponse<*> {
+    fun enableCoupon(@PathVariable couponId: Long, @LoginMember seller: Member): ApiResponse<*> {
         couponService.enable(couponId, seller)
         return ApiResponse.succeed()
+    }
+
+
+    @PostMapping("/{couponId}/download")
+    @Secured("ROLE_CUSTOMER")
+    fun downloadCoupon(@PathVariable couponId: Long, @LoginMember customer: Member): ApiResponse<MemberCouponResponse> {
+        return ApiResponse.succeed(couponService.downloadCoupon(couponId, customer))
     }
 
 }
