+++ src/main/kotlin/com/mojh/cms/coupon/controller/MemberCouponController.kt
@@ -1,13 +1,6 @@
 package com.mojh.cms.coupon.controller
 
-import com.mojh.cms.common.ApiResponse
-import com.mojh.cms.coupon.dto.MemberCouponResponse
 import com.mojh.cms.coupon.service.MemberCouponService
-import com.mojh.cms.member.entity.Member
-import com.mojh.cms.security.LoginMember
-import org.springframework.security.access.annotation.Secured
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
@@ -17,12 +10,4 @@ class MemberCouponController(
     val memberCouponService: MemberCouponService
 ) {
 
-
-    @PostMapping("/coupons/{couponId}/download")
-    @Secured("ROLE_CUSTOMER")
-    fun downloadCoupon(@PathVariable couponId: Long,
-                       @LoginMember customer: Member
-    ): ApiResponse<MemberCouponResponse> {
-        return ApiResponse.succeed(memberCouponService.downloadCoupon(couponId, customer))
-    }
 }
